@@ -1,100 +1,55 @@
-name: Habo
+name: habo
 description: Minimalistic habit tracker.
 
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+
+version: 2.0.0+1
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.18.1 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  table_calendar: ^2.3.3
-  sqflite: ^2.0.0+4
-  provider: ^6.0.0
-  shared_preferences: ^2.0.7
-  package_info: ^2.0.2
-  flutter_svg: ^0.22.0
-  flutter_local_notifications: ^8.2.0
-  url_launcher: ^6.0.10
-  awesome_dialog: ^2.0.0
-  just_audio: ^0.9.11
-  fl_chart: ^0.40.2
-  flutter_file_dialog: ^2.3.0
-  path_provider: ^2.0.7
-  loader_overlay: ^2.0.5
-  introduction_screen: ^2.1.0
+  google_fonts: ^3.0.1
+  provider: ^6.0.3
+  shared_preferences: ^2.0.15
+  table_calendar: ^3.0.7
+  awesome_dialog: ^3.0.2
+  loader_overlay: ^2.1.0
+  introduction_screen: ^3.0.2
+  flutter_svg: ^1.1.5
+  url_launcher: ^6.1.6
+  package_info_plus: ^1.4.3+1
+  sqflite: ^2.1.0+1
+  path: ^1.8.2
+  fl_chart: ^0.55.2
+  just_audio: ^0.9.30
+  flutter_file_dialog: ^2.3.2
+  path_provider: ^2.0.11
+  awesome_notifications: ^0.7.4+1
+  timezone: ^0.9.0
+  flutter_colorpicker: ^1.0.3
 
 dev_dependencies:
+  flutter_launcher_icons: "^0.10.0"
   flutter_test:
     sdk: flutter
+
+  flutter_lints: ^2.0.0
 
 flutter_icons:
   image_path: "assets/images/icon.png"
   android: true
   ios: true
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
 
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
 
-  # To add assets to your application, add an assets section, like this:
-  # assets:
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
   assets:
     - assets/
     - assets/images/
     - assets/images/onboard/
     - assets/sounds/
-
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
-  fonts:
-    - family: Nunito
-      fonts:
-        - asset: assets/fonts/Nunito/Nunito-Regular.ttf
-        - asset: assets/fonts/Nunito/Nunito-Bold.ttf
-          weight: 700
-        - asset: assets/fonts/Nunito/Nunito-Black.ttf
-          weight: 900
+    - assets/google_fonts/
