@@ -1,17 +1,51 @@
-import 'package:Habo/provider.dart';
-import 'package:Habo/screens/home_screen.dart';
-import 'package:Habo/screens/loading_screen.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:habo/habits/habits_manager.dart';
+import 'package:habo/navigation/app_router.dart';
+import 'package:habo/navigation/app_state_manager.dart';
+import 'package:habo/notifications.dart';
+import 'package:habo/settings/settings_manager.dart';
 import 'package:provider/provider.dart';
 
-void main() => runApp(Habo());
+void main() {
+  addLicenses();
+  runApp(
+    const Habo(),
+  );
+}
 
-class Habo extends StatelessWidget {
+class Habo extends StatefulWidget {
+  const Habo({Key? key}) : super(key: key);
+
+  @override
+  State<Habo> createState() => _HaboState();
+}
+
+class _HaboState extends State<Habo> {
+  final _appStateManager = AppStateManager();
+  final _settingsManager = SettingsManager();
+  final _habitManager = HabitsManager();
+  late AppRouter _appRouter;
+
+  @override
+  void initState() {
+    _settingsManager.initialize();
+    _habitManager.initialize();
+    initializeNotifications();
+    GoogleFonts.config.allowRuntimeFetching = false;
+    _appRouter = AppRouter(
+      appStateManager: _appStateManager,
+      settingsManager: _settingsManager,
+    );
+    super.initState();
+  }
+
   @override
   Widget build(BuildContext context) {
     SystemChrome.setSystemUIOverlayStyle(
-      SystemUiOverlayStyle(
+      const SystemUiOverlayStyle(
           statusBarColor: Colors.transparent,
           statusBarBrightness: Brightness.light),
     );
@@ -20,19 +54,36 @@ class Habo extends StatelessWidget {
     ]);
     return MultiProvider(
       providers: [
-        ChangeNotifierProvider(create: (_) => Bloc()),
+        ChangeNotifierProvider(
+          create: (context) => _appStateManager,
+        ),
+        ChangeNotifierProvider(
+          create: (context) => _settingsManager,
+        ),
+        ChangeNotifierProvider(
+          create: (context) => _habitManager,
+        ),
       ],
-      child: Consumer<Bloc>(
-        builder: (context, counter, _) {
-          final bloc = Provider.of<Bloc>(context);
-          return MaterialApp(
-            scaffoldMessengerKey: Provider.of<Bloc>(context).getScaffoldKey,
-            theme: Provider.of<Bloc>(context).getSettings.getLight,
-            darkTheme: Provider.of<Bloc>(context).getSettings.getDark,
-            home: !bloc.getDataLoaded ? LoadingScreen() : HomeScreen(),
-          );
-        },
-      ),
+      child: Consumer<SettingsManager>(builder: (context, counter, _) {
+        return MaterialApp(
+          title: 'Habo',
+          scaffoldMessengerKey:
+              Provider.of<HabitsManager>(context).getScaffoldKey,
+          theme: Provider.of<SettingsManager>(context).getLight,
+          darkTheme: Provider.of<SettingsManager>(context).getDark,
+          home: Router(
+            routerDelegate: _appRouter,
+            backButtonDispatcher: RootBackButtonDispatcher(),
+          ),
+        );
+      }),
     );
   }
+}
+
+void addLicenses() {
+  LicenseRegistry.addLicense(() async* {
+    final license = await rootBundle.loadString('assets/google_fonts/OFL.txt');
+    yield LicenseEntryWithLineBreaks(['google_fonts'], license);
+  });
 }
